backend/requirements.txt
#backend
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
python-dotenv
bcrypt

#AI
cmake
opencv-python
face_recognition
numpy
Pillow
